src/workers/worker_umps.py
import requests
from bs4 import BeautifulSoup
import pandas as pd
import time
import os
import logging
import sys

# IMPORT GOVERNANCE LAYERS
from src.governance import AgentContext, AgentContract, DiagnosisReport, ExecutionReport, DSIEStage
from src.base_worker import BaseWorker

class UmpsWorker(BaseWorker):
    """
    Worker for scraping Umpire Scorecards.
    """
    TARGET_URL = "https://umpscorecards.com/games/"

    def validate_html(self, content: bytes, name: str) -> bool:
        """Edge Data Integrity Check for HTML Content."""
        if not content:
            self.logger.error(f"Validation Failed: {name} is empty")
            return False
        if len(content) < 500: # Arbitrary small size check
            self.logger.warning(f"Validation Warning: {name} is suspiciously small ({len(content)} bytes)")
            return False
        return True

    def run(self) -> bool:
        self.logger.info("Checking the zone (Umpire Scorecards)...")
        
        try:
            self.logger.info(f"Scraping {self.TARGET_URL}...")
            resp = self.session.get(self.TARGET_URL, timeout=self.config.DEFAULT_TIMEOUT)
            
            if resp.status_code == 200:
                if self.validate_html(resp.content, "Scorecards Page"):
                    # Let's save the HTML regardless
                    filename = f"umpscorecards_games_{int(time.time())}.html"
                    blob = self.bucket.blob(f"results/mlb/{filename}")
                    blob.upload_from_string(resp.content, content_type='text/html')
                    self.logger.info(f"   -> Scorecards Secured (HTML).")
                    
                    # GENERATE EXECUTION REPORT
                    report = ExecutionReport(
                        stage=DSIEStage.EXECUTE,
                        subsystem=self.__class__.__name__,
                        change_summary="Umpire Scorecards Scrape",
                        primary_metric="bytes_secured",
                        metric_before=0.0,
                        metric_after=float(len(resp.content)),
                        observation_window_hours=0.01,
                        success=True,
                        notes=f"Secured: {len(resp.content)} bytes (HTML)"
                    )
                    self.file_report(report)
                    return True
                else:
                    self.logger.warning("   -> HTML validation failed.")
                    return False
            else:
                self.logger.error(f"Scrape failed: {resp.status_code}")
                return False

        except Exception as e:
            self.logger.error(f"Heist Failed: {e}")
            # Generate Failure Report
            fail_report = ExecutionReport(
                stage=DSIEStage.EXECUTE,
                subsystem=self.__class__.__name__,
                change_summary="Umpire Scorecards Scrape (FAILED)",
                primary_metric="bytes_secured",
                metric_before=0.0,
                metric_after=0.0,
                observation_window_hours=0.01,
                success=False,
                notes=str(e)
            )
            self.file_report(fail_report)
            return False

if __name__ == "__main__":
    logging.basicConfig(level=logging.INFO)
    
    # DEFINE THE AGENT
    contract = AgentContract(
        agent_id="acquisitions_officer_umps",
        human_readable_name="Acquisitions Officer (Umpire Scorecards)",
        autonomy_level=2
    )

    # TEST: RUN WITH DIAGNOSIS
    print("\n--- ATTEMPT: Running with valid paperwork ---")
    report = DiagnosisReport(
        problem_summary="Need umpire bias data",
        root_cause_hypothesis="Routine ingestion schedule",
        confidence=1.0
    )
    good_ctx = AgentContext(contract=contract, current_diagnosis=report)
    worker = UmpsWorker(good_ctx)
    worker.execute(ctx=good_ctx)
