requirements.txt
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
attrs==25.4.0
beautifulsoup4==4.14.2
bidict==0.23.1
CacheControl==0.14.4
cachetools==6.2.2
cattrs==25.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chickenstats==1.7.3.8.9.10.24
contourpy==1.3.3
cramjam==2.11.0
cryptography==46.0.3
cycler==0.12.1
fake-useragent==2.2.0
fastf1==3.7.0
fastparquet==2024.11.0
firebase_admin==7.1.0
fonttools==4.61.0
fsspec==2025.10.0
geopandas==1.1.1
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
joblib==1.5.2
kiwisolver==1.4.9
llvmlite==0.45.1
lxml==6.0.2
markdown-it-py==4.0.0
matplotlib==3.10.7
mdurl==0.1.2
msgpack==1.0.2
mypy_extensions==1.1.0
nfl_data_py==0.3.3
numba==0.62.1
numpy==1.26.4
nvidia-nccl-cu12==2.28.9
packaging==25.0
pandas==1.5.3
pandera==0.27.0
pillow==12.0.0
platformdirs==4.5.0
playwright==1.56.0
playwright-stealth==2.0.0
polars==1.35.2
polars-runtime-32==1.35.2
proto-plus==1.26.1
protobuf==6.33.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybaseball==2.2.7
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.1
pyogrio==0.12.1
pyparsing==3.2.5
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-socketio==5.15.0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.5
requests-cache==1.2.1
rich==14.2.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
shapely==2.1.2
signalrcore==0.9.5
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.8
threadpoolctl==3.6.0
timple==0.1.8
tqdm==4.67.1
typeguard==4.4.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
url-normalize==2.2.1
urllib3==2.5.0
webencodings==0.5.1
websocket-client==1.0.0
websockets==15.0.1
wsproto==1.3.2
xgboost==3.1.2
